requirements.txt
pandas
openpyxl
sentence-transformers
faiss-cpu
fastapi
uvicorn
torch
torchvision
torchaudio
python-multipart
requests
